+++ src/Pages/MainPage/MainPage.tsx
@@ -1,10 +1,16 @@
-import { FC, useEffect, useState, createContext, useContext, ReactNode } from 'react';
-
+import {
+  FC,
+  useEffect,
+  useState,
+  createContext,
+  useContext,
+  ReactNode,
+} from 'react';
 import { useLocation } from 'react-router-dom';
 import styled from 'styled-components';
 import Alarm from './components/Alarm';
 import NewProject from './components/NewProject';
-import Search from './components/Search';
+import Search from '../Layouts/Search';
 import BeforeLogin from '../Layouts/BeforeLogin';
 import Banner from './components/Banner';
 import ProjectList from './components/ProjectList';
@@ -31,179 +37,194 @@ const GlobalStyle = createGlobalStyle`
 `;
 
 const MainContainer = styled.div`
-    position: relative;
-    display: flex;
-    flex-direction: column;
-    margin: 0 auto;
-    max-width: 1284px;
-    box-sizing: border-box;
+  position: relative;
+  display: flex;
+  flex-direction: column;
+  margin: 0 auto;
+  max-width: 1284px;
+  box-sizing: border-box;
 `;
 
 function useQuery() {
-    const location = useLocation();
-    return new URLSearchParams(location.search);
+  const location = useLocation();
+  return new URLSearchParams(location.search);
 }
 
 type UserContextType = {
-    userData: MemberEntity | null;
-    setUserData: (userData: MemberEntity | null) => void;
+  userData: MemberEntity | null;
+  setUserData: (userData: MemberEntity | null) => void;
 };
 
 const UserContext = createContext<UserContextType | null>(null);
 
-export const UserProvider: React.FC<{ children: ReactNode }> = ({ children }) => {
-    const [userData, setUserData] = useState<MemberEntity | null>(null);
+export const UserProvider: React.FC<{ children: ReactNode }> = ({
+  children,
+}) => {
+  const [userData, setUserData] = useState<MemberEntity | null>(null);
 
-    return <UserContext.Provider value={{ userData, setUserData }}>{children}</UserContext.Provider>;
+  return (
+    <UserContext.Provider value={{ userData, setUserData }}>
+      {children}
+    </UserContext.Provider>
+  );
 };
 
 export const useUserContext = () => {
-    const context = useContext(UserContext);
-    if (context === null) throw new Error('useUserContext must be used within a UserProvider');
-    return context;
+  const context = useContext(UserContext);
+  if (context === null)
+    throw new Error('useUserContext must be used within a UserProvider');
+  return context;
 };
 
 const MainPage: FC = () => {
-    const { userData, setUserData } = useUserContext();
-    const query = useQuery();
-    const [isLoggedIn, setIsLoggedIn] = useState<boolean>(false);
-    const [projects, setProjects] = useState<ProjectEntity[]>([]);
-    const [searchResults, setSearchResults] = useState<ProjectEntity[]>([]);
+  const { userData, setUserData } = useUserContext();
+  const query = useQuery();
+  const [isLoggedIn, setIsLoggedIn] = useState<boolean>(false);
+  const [projects, setProjects] = useState<ProjectEntity[]>([]);
+  const [searchResults, setSearchResults] = useState<ProjectEntity[]>([]);
 
-    useEffect(() => {
-        const accessToken = query.get('access_token') || localStorage.getItem('access_token');
-        const refreshToken = query.get('refresh_token') || localStorage.getItem('refresh_token');
-        const memberId = query.get('member_id') || localStorage.getItem('member_id');
+  useEffect(() => {
+    const accessToken =
+      query.get('access_token') || localStorage.getItem('access_token');
+    const refreshToken =
+      query.get('refresh_token') || localStorage.getItem('refresh_token');
+    const memberId =
+      query.get('member_id') || localStorage.getItem('member_id');
 
-        if (accessToken) localStorage.setItem('access_token', accessToken);
-        if (refreshToken) localStorage.setItem('refresh_token', refreshToken);
-        if (memberId) localStorage.setItem('member_id', memberId);
+    if (accessToken) localStorage.setItem('access_token', accessToken);
+    if (refreshToken) localStorage.setItem('refresh_token', refreshToken);
+    if (memberId) localStorage.setItem('member_id', memberId);
 
-        if (accessToken && refreshToken) {
-            setIsLoggedIn(true);
-        } else {
-            setIsLoggedIn(false);
+    if (accessToken && refreshToken) {
+      setIsLoggedIn(true);
+    } else {
+      setIsLoggedIn(false);
+    }
+  }, []);
+
+  useEffect(() => {
+    const accessToken = localStorage.getItem('access_token');
+    const memberId = localStorage.getItem('member_id');
+    console.log(accessToken);
+    console.log(memberId);
+    const fetchUser = async () => {
+      try {
+        const response = await fetch(Http + `/v1/members/${memberId}`, {
+          method: 'GET',
+          headers: {
+            Accept: '*/*',
+            Authorization: `Bearer ${accessToken}`,
+          },
+        });
+        const data = await response.json();
+        console.log(data);
+        setUserData(data?.data);
+      } catch (error) {
+        console.error('Error fetching data:', error);
+      }
+    };
+    fetchUser();
+  }, [userData?.nickname, userData?.state]);
+
+  useEffect(() => {
+    const accessToken = localStorage.getItem('access_token');
+    const memberId = localStorage.getItem('member_id');
+    const fetchProjects = async () => {
+      const response = await fetch(
+        `${Http}/v1/projects/members/${memberId}?page=0&size=6`,
+        {
+          headers: {
+            Authorization: `Bearer ${accessToken}`,
+          },
         }
-    }, []);
-
-    useEffect(() => {
-        const accessToken = localStorage.getItem('access_token');
-        const memberId = localStorage.getItem('member_id');
-        console.log(accessToken);
-        console.log(memberId);
-        const fetchUser = async () => {
-            try {
-                const response = await fetch(Http + `/v1/members/${memberId}`, {
-                    method: 'GET',
-                    headers: {
-                        Accept: '*/*',
-                        Authorization: `Bearer ${accessToken}`,
-                    },
-                });
-                const data = await response.json();
-                console.log(data);
-                setUserData(data?.data);
-            } catch (error) {
-                console.error('Error fetching data:', error);
-            }
-        };
-        fetchUser();
-    }, [userData?.nickname, userData?.state]);
-
-    useEffect(() => {
-        const accessToken = localStorage.getItem('access_token');
-        const memberId = localStorage.getItem('member_id');
-        const fetchProjects = async () => {
-            const response = await fetch(`${Http}/v1/projects/members/${memberId}?page=0&size=6`, {
-                headers: {
-                    Authorization: `Bearer ${accessToken}`,
-                },
-            });
-            const data = await response.json();
-            console.log(data);
-            setProjects(data.data);
-            setSearchResults(data.data);
-        };
-
-        if (accessToken) fetchProjects();
-    }, []);
-
-    const handleSearch = (query: string) => {
-        const searchProjects = projects.filter((project) => project.title.toLowerCase().includes(query.toLowerCase()));
-        setSearchResults(searchProjects);
+      );
+      const data = await response.json();
+      console.log(data);
+      setProjects(data.data);
+      setSearchResults(data.data);
     };
 
-    return (
-        <>
-            <GlobalStyle />
+    if (accessToken) fetchProjects();
+  }, []);
 
-            {isLoggedIn ? <AfterLogin /> : <BeforeLogin />}
-            <Banner />
-            <div
-                style={{
-                    height: '126px',
-                }}
-            />
-            <MainContainer>
-                <div
-                    style={{
-                        display: 'flex',
-                        flexDirection: 'row',
-                        gap: '20px',
-                    }}
-                >
-                    <div
-                        style={{
-                            display: 'flex',
-                            flexDirection: 'column',
-                            alignItems: 'flex-start',
-                            gap: '14px',
-                        }}
-                    >
-                        <Profile />
-                        <Alarm />
-                    </div>
-                    <div
-                        style={{
-                            display: 'flex',
-                            flexDirection: 'column',
-                            alignItems: 'flex-start',
-                            gap: '24px',
-                        }}
-                    >
-                        <div
-                            style={{
-                                display: 'flex',
-                                flexDirection: 'column',
-                                alignItems: 'flex-start',
-                                gap: '21px',
-                            }}
-                        >
-                            <div
-                                style={{
-                                    display: 'flex',
-                                    flexDirection: 'row',
-                                    alignItems: 'flex-start',
-                                    gap: '8px',
-                                }}
-                            >
-                                <Search onSearch={handleSearch} />
-                                <NewProject />
-                            </div>
-                            <Pinned />
-                        </div>
-                        <ProjectList projects={searchResults} />
-                    </div>
-                </div>
-            </MainContainer>
-
-            <div
-                style={{
-                    height: '300px',
-                }}
-            />
-        </>
+  const handleSearch = (query: string) => {
+    const searchProjects = projects.filter((project) =>
+      project.title.toLowerCase().includes(query.toLowerCase())
     );
+    setSearchResults(searchProjects);
+  };
+
+  return (
+    <>
+      <GlobalStyle />
+
+      {isLoggedIn ? <AfterLogin /> : <BeforeLogin />}
+      <Banner />
+      <div
+        style={{
+          height: '126px',
+        }}
+      />
+      <MainContainer>
+        <div
+          style={{
+            display: 'flex',
+            flexDirection: 'row',
+            gap: '20px',
+          }}
+        >
+          <div
+            style={{
+              display: 'flex',
+              flexDirection: 'column',
+              alignItems: 'flex-start',
+              gap: '14px',
+            }}
+          >
+            <Profile />
+            <Alarm />
+          </div>
+          <div
+            style={{
+              display: 'flex',
+              flexDirection: 'column',
+              alignItems: 'flex-start',
+              gap: '24px',
+            }}
+          >
+            <div
+              style={{
+                display: 'flex',
+                flexDirection: 'column',
+                alignItems: 'flex-start',
+                gap: '21px',
+              }}
+            >
+              <div
+                style={{
+                  display: 'flex',
+                  flexDirection: 'row',
+                  alignItems: 'flex-start',
+                  gap: '8px',
+                }}
+              >
+                <Search onSearch={handleSearch} />
+                <NewProject />
+              </div>
+              <Pinned />
+            </div>
+            <ProjectList projects={searchResults} />
+          </div>
+        </div>
+      </MainContainer>
+
+      <div
+        style={{
+          height: '300px',
+        }}
+      />
+    </>
+  );
 };
 
 export default MainPage;
